feat: 콘솔 앱 구현
